requirements.txt
# Core Django
Django==5.2

# REST API support
djangorestframework==3.15.1

# PostgreSQL support
psycopg2-binary==2.9.9

djangorestframework-simplejw==5.5.0
